Move Settlement rating into separate component
