+++ crates/solver/src/lib.rs
@@ -12,6 +12,7 @@ pub mod orderbook;
 pub mod settlement;
 pub mod settlement_access_list;
 pub mod settlement_post_processing;
+pub mod settlement_rater;
 pub mod settlement_simulation;
 pub mod settlement_submission;
 pub mod solver;
